Finish Portfolio feature
