+++ src/Features/Portfolio/index.js
@@ -1,0 +1,3 @@
+import Portfolio from './Portfolio';
+
+export default Portfolio;
